Combined Single Card View JS Files
