+++ src/pages/single-card-view/single-card-view.html
@@ -4,7 +4,10 @@
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Single Card View</title>
+
     <link rel="stylesheet" href="./style.css">
+    <script src="single-card.js"></script>
+    <script src="single-card-manage.js"></script>
   </head>
   <body>
     <div class="card-container">
@@ -20,13 +23,11 @@
       <div class="carousel-wrapper">
         <button
           class="navigation prev-button"
-          onclick="prevCard()"
           title="Previous card"
         >
         </button>
         <button
           class="navigation next-button"
-          onclick="nextCard()"
           title="Next card"
         >
         </button>
